добавила создание визита

+++ Web/Models/VisitDetailsViewModel.cs
@@ -9,7 +9,7 @@ namespace Web.Models
     {
         public int Id { get; set; }
         public DateTime VisitData { get; set; }
-        public int OrderAmount { get; set; }
+        public double OrderAmount { get; set; }
         public string OrderStatus { get; set; }
     }
 }
